requirements.txt
flask
flask_socketio
jsonify
GitPython
transformers
langchain
langchain_huggingface
accelerate
torch
GitPython
requests
